test(widget): move textarea bounds tests to separate integration test file
